move quickinputservice to own file
