+++ src/vs/editor/standalone/browser/quickInput/standaloneQuickInputService.ts
@@ -15,7 +15,8 @@ import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { IAccessibilityService } from 'vs/platform/accessibility/common/accessibility';
 import { EditorScopedLayoutService } from 'vs/editor/standalone/browser/standaloneLayoutService';
 import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
-import { QuickInputService, IQuickInputControllerHost, QuickInputController } from 'vs/platform/quickinput/browser/quickInput';
+import { IQuickInputControllerHost, QuickInputController } from 'vs/platform/quickinput/browser/quickInput';
+import { QuickInputService } from 'vs/platform/quickinput/browser/quickInputService';
 import { once } from 'vs/base/common/functional';
 import { IQuickAccessController } from 'vs/platform/quickinput/common/quickAccess';
